Move libraries for api.ResourceList to a utility pkg
