+++ pkg/api/util/resources/OWNERS
@@ -1,0 +1,3 @@
+assignees:
+ - derekwaynecarr
+ - vishh
