+++ pkg/quota/evaluator/core/services_test.go
@@ -21,8 +21,8 @@ import (
 
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/resource"
+	"k8s.io/kubernetes/pkg/api/util/resources"
 	"k8s.io/kubernetes/pkg/client/clientset_generated/release_1_5/fake"
-	"k8s.io/kubernetes/pkg/quota"
 )
 
 func TestServiceEvaluatorMatchesResources(t *testing.T) {
@@ -37,12 +37,12 @@ func TestServiceEvaluatorMatchesResources(t *testing.T) {
 		api.ResourceServicesLoadBalancers,
 	}
 	// but we only match these...
-	expected := quota.ToSet([]api.ResourceName{
+	expected := resources.ToSet([]api.ResourceName{
 		api.ResourceServices,
 		api.ResourceServicesNodePorts,
 		api.ResourceServicesLoadBalancers,
 	})
-	actual := quota.ToSet(evaluator.MatchingResources(input))
+	actual := resources.ToSet(evaluator.MatchingResources(input))
 	if !expected.Equal(actual) {
 		t.Errorf("expected: %v, actual: %v", expected, actual)
 	}
@@ -122,7 +122,7 @@ func TestServiceEvaluatorUsage(t *testing.T) {
 		if err != nil {
 			t.Errorf("%s unexpected error: %v", testName, err)
 		}
-		if !quota.Equals(testCase.usage, actual) {
+		if !resources.Equals(testCase.usage, actual) {
 			t.Errorf("%s expected: %v, actual: %v", testName, testCase.usage, actual)
 		}
 	}
