+++ plugin/pkg/admission/resourcequota/BUILD
@@ -21,6 +21,7 @@ go_library(
         "//pkg/admission:go_default_library",
         "//pkg/api:go_default_library",
         "//pkg/api/meta:go_default_library",
+        "//pkg/api/util/resources:go_default_library",
         "//pkg/api/v1:go_default_library",
         "//pkg/client/cache:go_default_library",
         "//pkg/client/clientset_generated/internalclientset:go_default_library",
