+++ test/test_owners.csv
@@ -63,7 +63,7 @@ Density create a batch of pods latency/resource should be within limit when crea
 Density create a batch of pods with higher API QPS latency/resource should be within limit when create * pods with * interval (QPS *),jlowdermilk,1
 Density create a sequence of pods latency/resource should be within limit when create * pods with * background pods,wojtek-t,1
 Density should allow running maximum capacity pods on nodes,smarterclayton,1
-Density should allow starting * pods per node using *,derekwaynecarr,0
+Density should allow starting * pods per node using * with * secrets,derekwaynecarr,0
 Deployment RecreateDeployment should delete old pods and create new ones,pwittrock,0
 Deployment RollingUpdateDeployment should delete old pods and create new ones,pwittrock,0
 Deployment RollingUpdateDeployment should scale up and down in the right order,pwittrock,0
@@ -275,7 +275,7 @@ KubeletManagedEtcHosts should test kubelet managed /etc/hosts file,Random-Liu,1
 Kubernetes Dashboard should check that the kubernetes-dashboard instance is alive,wonderfly,0
 LimitRange should create a LimitRange with defaults and ensure pod has those defaults applied.,cjcullen,1
 Liveness liveness pods should be automatically restarted,derekwaynecarr,0
-Load capacity should be able to handle * pods per node *,derekwaynecarr,0
+Load capacity should be able to handle * pods per node * with * secrets,derekwaynecarr,0
 Loadbalancing: L7 GCE shoud create ingress with given static-ip,derekwaynecarr,0
 Loadbalancing: L7 GCE should conform to Ingress spec,derekwaynecarr,0
 Loadbalancing: L7 Nginx should conform to Ingress spec,ncdc,1
@@ -303,7 +303,7 @@ Network Partition Pods should return to running and ready state after network pa
 Network Partition should come back up if node goes down,foxish,0
 Network Partition should create new pods when node is partitioned,foxish,0
 Network Partition should eagerly create replacement pod during network partition when termination grace is non-zero,foxish,0
-Network Partition should not reschedule pets if there is a network partition,foxish,0
+Network Partition should not reschedule stateful pods if there is a network partition,derekwaynecarr,0
 Network should set TCP CLOSE_WAIT timeout,bowei,0
 Networking Granular Checks: Pods should function for intra-pod communication: http,stts,0
 Networking Granular Checks: Pods should function for intra-pod communication: udp,freehan,0
@@ -465,9 +465,9 @@ SimpleMount should be able to mount an emptydir on a container,derekwaynecarr,0
 "Staging client repo client should create pods, delete pods, watch pods",jbeda,1
 Stateful Set recreate should recreate evicted statefulset,derekwaynecarr,0
 StatefulSet Basic StatefulSet functionality Scaling down before scale up is finished should wait until current pod will be running and ready before it will be removed,derekwaynecarr,0
-StatefulSet Basic StatefulSet functionality Scaling should happen in predictable order and halt if any pet is unhealthy,derekwaynecarr,0
+StatefulSet Basic StatefulSet functionality Scaling should happen in predictable order and halt if any stateful pod is unhealthy,derekwaynecarr,0
 StatefulSet Basic StatefulSet functionality should allow template updates,derekwaynecarr,0
-StatefulSet Basic StatefulSet functionality should handle healthy pet restarts during scale,kevin-wangzefeng,1
+StatefulSet Basic StatefulSet functionality should handle healthy stateful pod restarts during scale,derekwaynecarr,0
 StatefulSet Basic StatefulSet functionality should provide basic identity,bprashanth,1
 StatefulSet Deploy clustered applications should creating a working CockroachDB cluster,derekwaynecarr,0
 StatefulSet Deploy clustered applications should creating a working mysql cluster,yujuhong,1
@@ -516,6 +516,7 @@ k8s.io/kubernetes/cmd/kubeadm/app/preflight,apprenda,0
 k8s.io/kubernetes/cmd/kubeadm/app/util,krousey,1
 k8s.io/kubernetes/cmd/kubeadm/test,pipejakob,0
 k8s.io/kubernetes/cmd/kubelet/app,derekwaynecarr,0
+k8s.io/kubernetes/cmd/kubernetes-discovery/pkg/apiserver,derekwaynecarr,0
 k8s.io/kubernetes/cmd/libs/go2idl/client-gen/types,caesarxuchao,0
 k8s.io/kubernetes/cmd/libs/go2idl/go-to-protobuf/protobuf,smarterclayton,0
 k8s.io/kubernetes/cmd/libs/go2idl/openapi-gen/generators,davidopp,1
@@ -526,6 +527,7 @@ k8s.io/kubernetes/federation/apis/federation/validation,nikhiljindal,0
 k8s.io/kubernetes/federation/cmd/federation-controller-manager/app,kzwang,0
 k8s.io/kubernetes/federation/pkg/dnsprovider,sttts,1
 k8s.io/kubernetes/federation/pkg/dnsprovider/providers/aws/route53,cjcullen,1
+k8s.io/kubernetes/federation/pkg/dnsprovider/providers/coredns,derekwaynecarr,0
 k8s.io/kubernetes/federation/pkg/dnsprovider/providers/google/clouddns,jsafrane,1
 k8s.io/kubernetes/federation/pkg/federation-controller/cluster,nikhiljindal,0
 k8s.io/kubernetes/federation/pkg/federation-controller/configmap,mwielgus,0
@@ -556,6 +558,7 @@ k8s.io/kubernetes/pkg/api/resource,smarterclayton,1
 k8s.io/kubernetes/pkg/api/service,spxtr,1
 k8s.io/kubernetes/pkg/api/testapi,caesarxuchao,1
 k8s.io/kubernetes/pkg/api/util,derekwaynecarr,0
+k8s.io/kubernetes/pkg/api/util/resources,derekwaynecarr,0
 k8s.io/kubernetes/pkg/api/v1,derekwaynecarr,0
 k8s.io/kubernetes/pkg/api/v1/endpoints,derekwaynecarr,0
 k8s.io/kubernetes/pkg/api/v1/pod,derekwaynecarr,0
@@ -578,8 +581,6 @@ k8s.io/kubernetes/pkg/apis/extensions,bgrant0607,1
 k8s.io/kubernetes/pkg/apis/extensions/v1beta1,madhusudancs,1
 k8s.io/kubernetes/pkg/apis/extensions/validation,nikhiljindal,1
 k8s.io/kubernetes/pkg/apis/meta/v1,sttts,0
-k8s.io/kubernetes/pkg/apis/meta/v1/validation,derekwaynecarr,0
-k8s.io/kubernetes/pkg/apis/meta/v1/validation,jszczepkowski,1
 k8s.io/kubernetes/pkg/apis/meta/v1/unstructured,smarterclayton,0
 k8s.io/kubernetes/pkg/apis/meta/v1/validation,smarterclayton,0
 k8s.io/kubernetes/pkg/apis/policy/validation,deads2k,1
@@ -651,6 +652,7 @@ k8s.io/kubernetes/pkg/conversion,ixdy,1
 k8s.io/kubernetes/pkg/conversion/queryparams,caesarxuchao,1
 k8s.io/kubernetes/pkg/credentialprovider,justinsb,1
 k8s.io/kubernetes/pkg/credentialprovider/aws,zmerlynn,1
+k8s.io/kubernetes/pkg/credentialprovider/azure,derekwaynecarr,0
 k8s.io/kubernetes/pkg/credentialprovider/gcp,mml,1
 k8s.io/kubernetes/pkg/dns,derekwaynecarr,0
 k8s.io/kubernetes/pkg/dns/config,derekwaynecarr,0
@@ -663,7 +665,6 @@ k8s.io/kubernetes/pkg/genericapiserver/authorizer,lavalamp,1
 k8s.io/kubernetes/pkg/genericapiserver/filters,dchen1107,1
 k8s.io/kubernetes/pkg/genericapiserver/mux,spxtr,1
 k8s.io/kubernetes/pkg/genericapiserver/openapi,davidopp,1
-k8s.io/kubernetes/pkg/genericapiserver/routes,vishh,1
 k8s.io/kubernetes/pkg/healthz,thockin,1
 k8s.io/kubernetes/pkg/httplog,mtaufen,1
 k8s.io/kubernetes/pkg/kubectl,madhusudancs,1
@@ -723,7 +724,6 @@ k8s.io/kubernetes/pkg/proxy/healthcheck,derekwaynecarr,0
 k8s.io/kubernetes/pkg/proxy/iptables,freehan,0
 k8s.io/kubernetes/pkg/proxy/userspace,luxas,1
 k8s.io/kubernetes/pkg/proxy/winuserspace,jbhurat,0
-k8s.io/kubernetes/pkg/quota,sttts,1
 k8s.io/kubernetes/pkg/quota/evaluator/core,yifan-gu,1
 k8s.io/kubernetes/pkg/registry/apps/petset,kevin-wangzefeng,1
 k8s.io/kubernetes/pkg/registry/apps/petset/etcd,piosz,1
@@ -955,6 +955,7 @@ k8s.io/kubernetes/test/integration/auth,jbeda,1
 k8s.io/kubernetes/test/integration/client,Q-Lee,1
 k8s.io/kubernetes/test/integration/configmap,Q-Lee,1
 k8s.io/kubernetes/test/integration/discoverysummarizer,fabioy,1
+k8s.io/kubernetes/test/integration/evictions,derekwaynecarr,0
 k8s.io/kubernetes/test/integration/examples,maisem,1
 k8s.io/kubernetes/test/integration/federation,derekwaynecarr,0
 k8s.io/kubernetes/test/integration/garbagecollector,jlowdermilk,1
